+++ src/main/java/com/deliveryfood/service/IMenuService.java
@@ -1,0 +1,14 @@
+package com.deliveryfood.service;
+
+import com.deliveryfood.dto.MenuDto;
+import com.deliveryfood.model.MenuInput;
+
+import java.util.List;
+
+public interface IMenuService {
+
+    void createMenuById(MenuInput menuInput);
+    MenuDto findMenus(MenuInput menuInput);
+    List<MenuDto> findMenuById(MenuInput menuInput);
+    void modifyMenuById(MenuInput menuInput);
+}
