+++ src/main/java/com/deliveryfood/service/RestaurantService.java
@@ -1,16 +1,59 @@
 package com.deliveryfood.service;
 
 import com.deliveryfood.dto.RestaurantDto;
+import com.deliveryfood.mapper.RestaurantMapper;
 import com.deliveryfood.model.RestaurantInput;
-import com.deliveryfood.model.UserInput;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-public interface RestaurantService {
+@RequiredArgsConstructor
+@Service
+public class RestaurantService implements IRestaurantService {
 
-    void signin(RestaurantInput restaurantInput);
-    void signout(RestaurantInput restaurantInput);
-    List<RestaurantDto> findUsers();
-    RestaurantDto findUserById(RestaurantInput restaurantInput);
-    void modifyUserById(RestaurantInput restaurantInput);
+    private final RestaurantMapper restaurantMapper;
+
+    @Override
+    public void signin(RestaurantInput restaurantInput) {
+        RestaurantDto restaurantDto = RestaurantDto.builder()
+                .restaurantId(restaurantInput.getRestaurantId())
+                .userId(restaurantInput.getUserId())
+                .name(restaurantInput.getName())
+                .build();
+        restaurantMapper.signin(restaurantDto);
+    }
+
+    @Override
+    public void signout(RestaurantInput restaurantInput) {
+        RestaurantDto restaurantDto = RestaurantDto.builder()
+                .restaurantId(restaurantInput.getRestaurantId())
+                .build();
+        restaurantMapper.signout(restaurantDto);
+    }
+
+    @Override
+    public List<RestaurantDto> findUsers() {
+        return restaurantMapper.findUsers();
+    }
+
+    @Override
+    public RestaurantDto findUserById(RestaurantInput restaurantInput) {
+        RestaurantDto restaurantDto = RestaurantDto.builder()
+                .restaurantId(restaurantInput.getRestaurantId())
+                .userId(restaurantInput.getUserId())
+                .build();
+        return restaurantMapper.findUserById(restaurantDto);
+    }
+
+    @Override
+    public void modifyUserById(RestaurantInput restaurantInput) {
+        RestaurantDto restaurantDto = RestaurantDto.builder()
+                .restaurantId(restaurantInput.getRestaurantId())
+                .name(restaurantInput.getName())
+                .build();
+        restaurantMapper.modifyUserById(restaurantDto);
+    }
+
+
 }
