+++ src/main/java/com/deliveryfood/service/UserService.java
@@ -1,6 +1,6 @@
 package com.deliveryfood.service;
 
-import com.deliveryfood.Util.MemberSession;
+import com.deliveryfood.util.MemberSession;
 import com.deliveryfood.dao.MemberDao;
 import com.deliveryfood.dao.UserDao;
 import com.deliveryfood.dto.MemberDto;
